Rewrites the MetricOutput struct to InstanceMetric
Did this to improve the raw json output structure. Also for now, it is
conforming strictly to json parsable output from the metrics endpoint.
If the metrics endpoint doesn't provide a json representation, this
plugin will fail. Will add support for other formats if there is
feedback to support it.